# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ActionResult": {
        "additionalProperties": false,
        "properties": {
          "elapsed_ms": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Ms"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "output": {
            "additionalProperties": true,
            "title": "Output",
            "type": "object"
          }
        },
        "title": "ActionResult",
        "type": "object"
      },
      "ArtifactInfo": {
        "additionalProperties": false,
        "properties": {
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "created_by_call_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Call Id"
          },
          "kind": {
            "enum": [
              "file",
              "directory"
            ],
            "title": "Kind",
            "type": "string"
          },
          "modified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified At"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "retention_class": {
            "default": "ephemeral",
            "enum": [
              "ephemeral",
              "persistent",
              "debug",
              "trace"
            ],
            "title": "Retention Class",
            "type": "string"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "size_bytes": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "uri": {
            "title": "Uri",
            "type": "string"
          }
        },
        "required": [
          "path",
          "uri",
          "kind"
        ],
        "title": "ArtifactInfo",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HoldRequest": {
        "additionalProperties": false,
        "properties": {
          "actions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Actions",
            "type": "array"
          },
          "duration_ms": {
            "anyOf": [
              {
                "maximum": 60000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "key": {
            "title": "Key",
            "type": "string"
          }
        },
        "required": [
          "key"
        ],
        "title": "HoldRequest",
        "type": "object"
      },
...
      "HotkeyRequest": {
        "additionalProperties": false,
        "properties": {
          "duration_ms": {
            "default": 0,
            "maximum": 60000.0,
            "minimum": 0.0,
            "title": "Duration Ms",
            "type": "integer"
          },
          "keys": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Keys",
            "type": "array"
          }
        },
        "required": [
          "keys"
        ],
        "title": "HotkeyRequest",
        "type": "object"
      },
...
      "KeyRequest": {
        "additionalProperties": false,
        "properties": {
          "duration_ms": {
            "default": 0,
            "maximum": 60000.0,
            "minimum": 0.0,
            "title": "Duration Ms",
            "type": "integer"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "modifiers": {
            "items": {
              "type": "string"
            },
            "title": "Modifiers",
            "type": "array"
          }
        },
        "required": [
          "key"
        ],
        "title": "KeyRequest",
        "type": "object"
      },
...
      "TypeRequest": {
        "additionalProperties": false,
        "properties": {
          "delay_ms": {
            "default": 10,
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Delay Ms",
            "type": "integer"
          },
          "method": {
            "default": "auto",
            "enum": [
              "auto",
              "xdotool",
              "clipboard"
            ],
            "title": "Method",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "TypeRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "modal-computer-use daemon",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/artifacts": {
      "get": {
        "operationId": "list_artifacts_v1_artifacts_get",
        "parameters": [
          {
            "in": "query",
            "name": "prefix",
            "required": false,
            "schema": {
              "default": "",
              "title": "Prefix",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArtifactInfo"
                  },
                  "title": "Response List Artifacts V1 Artifacts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Artifacts"
      }
    },
    "/v1/artifacts/manifest": {
      "get": {
        "operationId": "manifest_v1_artifacts_manifest_get",
        "parameters": [
          {
            "in": "query",
            "name": "prefix",
            "required": false,
            "schema": {
              "default": "",
              "title": "Prefix",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArtifactInfo"
                  },
                  "title": "Response Manifest V1 Artifacts Manifest Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Manifest"
      }
    },
...
    "/v1/artifacts/{path}": {
      "delete": {
        "operationId": "delete_artifact_v1_artifacts__path__delete",
        "parameters": [
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Artifact V1 Artifacts  Path  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Artifact"
      },
      "get": {
        "operationId": "read_artifact_v1_artifacts__path__get",
        "parameters": [
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Artifact"
      },
      "put": {
        "operationId": "write_artifact_v1_artifacts__path__put",
        "parameters": [
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Write Artifact"
      }
    },
...
    "/v1/keyboard/hold": {
      "post": {
        "operationId": "hold_v1_keyboard_hold_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HoldRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Hold"
      }
    },
    "/v1/keyboard/hotkey": {
      "post": {
        "operationId": "hotkey_v1_keyboard_hotkey_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotkeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Hotkey"
      }
    },
...
    "/v1/keyboard/press": {
      "post": {
        "operationId": "press_v1_keyboard_press_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Press"
      }
    },
    "/v1/keyboard/type": {
      "post": {
        "operationId": "keyboard_type_v1_keyboard_type_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TypeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Keyboard Type"
      }
    },
...
  }
}